+++ examples/grpc/server.cc
@@ -1,11 +1,11 @@
 #ifdef BAZEL_BUILD
-#include "examples/grpc/protos/messages.grpc.pb.h"
+#  include "examples/grpc/protos/messages.grpc.pb.h"
 #else
-#include "messages.grpc.pb.h"
+#  include "messages.grpc.pb.h"
 #endif
 
 #include "opentelemetry/trace/context.h"
-#include "opentelemetry/trace/semantic_conventions.h"
+#include "opentelemetry/trace/experimental_semantic_conventions.h"
 #include "opentelemetry/trace/span_context_kv_iterable_view.h"
 #include "tracer_common.h"
 
@@ -16,10 +16,10 @@
 
 #include <chrono>
 #include <fstream>
+#include <map>
 #include <sstream>
 #include <string>
 #include <thread>
-#include <map>
 
 using grpc::Server;
 using grpc::ServerBuilder;
@@ -31,7 +31,7 @@ using grpc_example::Greeter;
 using grpc_example::GreetRequest;
 using grpc_example::GreetResponse;
 
-using Span = opentelemetry::trace::Span;
+using Span        = opentelemetry::trace::Span;
 using SpanContext = opentelemetry::trace::SpanContext;
 using namespace opentelemetry::trace;
 
@@ -44,7 +44,8 @@ public:
                const GreetRequest *request,
                GreetResponse *response) override
   {
-    for(  auto elem: context->client_metadata()) {
+    for (auto elem : context->client_metadata())
+    {
       std::cout << "ELEM: " << elem.first << " " << elem.second << "\n";
     }
 
@@ -61,13 +62,13 @@ public:
     options.parent   = opentelemetry::trace::GetSpan(new_context)->GetContext();
 
     std::string span_name = "GreeterService/Greet";
-    auto span             = get_tracer("grpc")
-                    ->StartSpan(span_name,
-                                {{OTEL_CPP_GET_ATTR(AttrRpcSystem), "grpc"},
-                                 {OTEL_CPP_GET_ATTR(AttrRpcService), "GreeterService"},
-                                 {OTEL_CPP_GET_ATTR(AttrRpcMethod), "Greet"},
-                                 {OTEL_CPP_GET_ATTR(AttrRpcGrpcStatusCode), 0}},
-                                options);
+    auto span =
+        get_tracer("grpc")->StartSpan(span_name,
+                                      {{OTEL_CPP_GET_ATTR(AttrRpcSystem), "grpc"},
+                                       {OTEL_CPP_GET_ATTR(AttrRpcService), "GreeterService"},
+                                       {OTEL_CPP_GET_ATTR(AttrRpcMethod), "Greet"},
+                                       {OTEL_CPP_GET_ATTR(AttrRpcGrpcStatusCode), 0}},
+                                      options);
     auto scope = get_tracer("grpc")->WithActiveSpan(span);
 
     // Fetch and parse whatever HTTP headers we can from the gRPC request.
