+++ exporters/zipkin/src/recordable.cc
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "opentelemetry/exporters/zipkin/recordable.h"
-#include "opentelemetry/sdk/resource/semantic_conventions.h"
+#include "opentelemetry/sdk/resource/experimental_semantic_conventions.h"
 
 #include <map>
 #include <string>
